pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-context-builder"
version = "0.1.0"
description = "Autonomous context generation for agents: compact bootstrap, workspace triage, topic index"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DataCivicLab", email = "dev@dataciviclab.org"}
]
keywords = ["agent", "context", "github", "triage"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
agent-context = "agent_context_builder.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_context_builder"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/agent_context_builder"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = false
